requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
ddgs>=0.5.0
